+++ examples/loaded_models.py
@@ -14,12 +14,12 @@ from imagebaker.models.base_model import BaseDetectionModel
 from imagebaker import logger
 
 # Import models from the examples folder
-from examples.rtdetr_v2 import RTDetrModelConfig, RTDetrDetectionModel
-from examples.segmentation import (
-    YoloSegmentationModel,
-    YoloSegmentationModelConfig,
-)
-from examples.sam_model import SegmentAnythingModel, SAMModelConfig
+# from examples.rtdetr_v2 import RTDetrModelConfig, RTDetrDetectionModel
+# from examples.segmentation import (
+#     YoloSegmentationModel,
+#     YoloSegmentationModelConfig,
+# )
+# from examples.sam_model import SegmentAnythingModel, SAMModelConfig
 
 
 class ClassificationModel(BaseClassificationModel):
@@ -38,22 +38,22 @@ class DetectionModel(BaseDetectionModel):
         return [get_dummy_prediction_result(self.config.model_type)]
 
 
-detector = RTDetrDetectionModel(RTDetrModelConfig())
+# detector = RTDetrDetectionModel(RTDetrModelConfig())
 
-classification = ClassificationModel(
-    DefaultModelConfig(model_type=ModelType.CLASSIFICATION)
-)
-segmentation = YoloSegmentationModel(YoloSegmentationModelConfig())
-prompt = SegmentAnythingModel(SAMModelConfig())
+# classification = ClassificationModel(
+#     DefaultModelConfig(model_type=ModelType.CLASSIFICATION)
+# )
+# segmentation = YoloSegmentationModel(YoloSegmentationModelConfig())
+# prompt = SegmentAnythingModel(SAMModelConfig())
 dummy_detector = DetectionModel(DefaultModelConfig(model_type=ModelType.DETECTION))
 
 
 LOADED_MODELS = {
     "DummyDetectionModel": dummy_detector,
-    "PromptModel": prompt,
-    "SegmentationModel": segmentation,
-    "RTDetrV2": detector,
-    "ClassificationModel": classification,
+    # "PromptModel": prompt,
+    # "SegmentationModel": segmentation,
+    # "RTDetrV2": detector,
+    # "ClassificationModel": classification,
 }
 
 logger.info(f"Loaded models: {LOADED_MODELS}")