+++ imagebaker/__init__.py
@@ -2,4 +2,4 @@ from loguru import logger  # noqa
 
 logger.info("imagebaker package loaded with loguru logger.")
 
-__version__ = "0.0.48"  # noqa
+# __version__ = "0.0.49"  # noqa