+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## April 2 2025
+Worked on [#3](https://github.com/q-viper/image-baker/issues/3), [#4](https://github.com/q-viper/image-baker/issues/4), [#5](https://github.com/q-viper/image-baker/issues/5).
+
+## Fixed
+* Handling of point movement in polygon.
+* Can control the recursive image laoding or partial from config.
+
+## Added
+* Double click on edge of polygon will add a point.
+* Double click on a polygon point will remove that point.
+* Can move point polygon around.
+
+## Removed
+* Double click to select unselect annotation.
+
 ## March 24 2025
 ### Added
 * Tested with Segmentation, Detection and Prompt models.